Add files via upload
This project includes detailed analysis of the healthcare insurance dataset in both SQL and Python, along with the documentation work (writeup file). An ML ode is built to analyse and predict the best fitting model to answer the questions like which tier hospital is the best pick amongst the lot, etc. 
Along with showcasing the coding spectrum of the project there are dashboards included from Tableau to visually explore the data some more and generate meaningful stories to convey the exploration details discovered from this project.